+++ CustomerData/Fpc.php
@@ -11,7 +11,7 @@ use Magento\Framework\UrlInterface;
 /**
  * Class Fpc
  *
- * @version 1.0.4
+ * @version 1.0.5
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\CustomerData
@@ -76,7 +76,7 @@ class Fpc implements SectionSourceInterface
                 'eventName' => 'AddToCart',
                 'eventAdditional' => [
                     'productId' => $this->_retailRocketHelper->getSession()->getAddToCart(),
-                    'stockId' => $this->_retailRocketHelper->isStockIdEnabled() ? $this->_retailRocketHelper->getCurrentWebsiteCode() : null
+                    'stockId' => $this->_retailRocketHelper->isStockIdEnabled() ? $this->_retailRocketHelper->getCurrentStoreCode() : null
                 ],
             ];
         }