+++ Model/Config/Source/RootCategoryId.php
@@ -10,7 +10,7 @@ use Magento\Framework\Option\ArrayInterface;
 /**
  * Class RootCategoryId
  *
- * @version 1.0.4
+ * @version 1.0.5
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\Model\Config\Source