+++ Observer/CustomerLogin.php
@@ -12,7 +12,7 @@ use Magento\Customer\Model\Customer;
 /**
  * Class CustomerLogin
  *
- * @version 1.0.4
+ * @version 1.0.5
  * @author Improntus <http://www.improntus.com> - Ecommerce done right
  * @copyright Copyright (c) 2020 Improntus
  * @package Improntus\RetailRocket\Observer